implement base entities updateable/deleteable by creators

+++ src/main/java/com/vaszily/WorkoutPlanner/service/entities/imp/ExerciseService.java
@@ -1,5 +1,6 @@
 package com.vaszily.WorkoutPlanner.service.entities.imp;
 
+import com.vaszily.WorkoutPlanner.exception.MissingAuthorityException;
 import com.vaszily.WorkoutPlanner.exception.ReferencedEntityException;
 import com.vaszily.WorkoutPlanner.model.Exercise;
 import com.vaszily.WorkoutPlanner.repositories.ExerciseRepo;
@@ -8,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityNotFoundException;
+import java.security.Principal;
 import java.util.List;
 
 @Service
@@ -39,8 +41,10 @@ public class ExerciseService implements EntityService<Exercise> {
     }
 
     @Override
-    public Exercise update(Long id, Exercise exercise) {
+    public Exercise update(Long id, Exercise exercise, Principal principal) {
         Exercise toUpdate = exerciseRepo.findById(id).orElseThrow(EntityNotFoundException::new);
+        if(!toUpdate.getCreatedBy().equals(principal.getName()))
+            throw new MissingAuthorityException("Can't update this exercise! Different creator!");
         toUpdate.update(exercise);
         return exerciseRepo.save(toUpdate);
 
@@ -48,10 +52,12 @@ public class ExerciseService implements EntityService<Exercise> {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id, Principal principal) {
         Exercise toDelete = exerciseRepo.findById(id).orElseThrow(EntityNotFoundException::new);
         if(toDelete.getExerciseWrappers().size()>0)
             throw new ReferencedEntityException("This excersice used by at least a exercise wrapper!");
+        if(!toDelete.getCreatedBy().equals(principal.getName()))
+            throw new MissingAuthorityException("Can't delete this exercise! Different creator!");
         exerciseRepo.delete(toDelete);
     }
 }
